Refactored navbar and groups #111
